circuit/builder: renamed module word64 to keccak_types
